chore: move pages to pages/app-pages [YTFRONT-5069]
